move bottom sheet to player component
