+++ frontend-v3/src/app/footer/footer.component.html
@@ -1,8 +1,3 @@
 <div fxLayout="column" fxLayoutAlign="center center">
-    <div class="footer" fxLayout="row" fxLayoutAlign="space-evenly center">    
-        <mat-slider (input)="onSliderChange($event)" step="0.005" min="0" max="100" class="slider">
-            <input matSliderThumb [(ngModel)]="sliderValue"/>
-        </mat-slider>
-        <button mat-raised-button (click)="openBottomSheet()">open</button>
-    </div>
+    <div class="footer" fxLayout="row" fxLayoutAlign="space-evenly center"></div>
 </div>
