+++ frontend-v3/src/app/footer/footer.component.ts
@@ -1,6 +1,4 @@
 import { Component, OnInit } from '@angular/core';
-import {MatListModule} from '@angular/material/list';
-import { MatBottomSheet, MatBottomSheetModule, MatBottomSheetRef } from '@angular/material/bottom-sheet';
 
 
 import { ApiService } from '../api-service';
@@ -13,41 +11,9 @@ import { ApiService } from '../api-service';
 })
 
 export class FooterComponent implements OnInit {
-  lastReleaseString: string = '';
-  sliderValue: number = 0;
-
-  constructor(private apiService: ApiService, private _bottomSheet: MatBottomSheet) { }
+  constructor(private apiService: ApiService) { }
 
   ngOnInit() { }
-
-  onSliderChange(event: Event) {
-    
-  }
-
-  openBottomSheet(): void {
-    this._bottomSheet.open(BottomSheetOverviewExampleSheet);
-  }
 }
 
-@Component({
-  selector: 'app-footer-bottom-sheet',
-  styleUrl: './footer.component.css',
-  imports: [MatListModule],
-  standalone: true,
-  template: `
-  <mat-nav-list>
-    <a href="https://keep.google.com/" mat-list-item>
-        <span matListItemTitle>Google Keep</span>
-        <span matLine>Add to a note</span>
-    </a>
-  </mat-nav-list>
-  `
-})
-export class BottomSheetOverviewExampleSheet {
-  constructor(private _bottomSheetRef: MatBottomSheetRef<BottomSheetOverviewExampleSheet>) {}
 
-  openLink(event: MouseEvent): void {
-    this._bottomSheetRef.dismiss();
-    event.preventDefault();
-  }
-}
