+++ frontend-v3/src/app/player/player.component.html
@@ -32,6 +32,11 @@
                 <p class="duration-time-text">&nbsp;{{ musicLength }}</p>
             </div>
         </div>
-        <div fxLayout="row" fxLayoutAlign="space-evenly center"></div>
+        <div fxLayout="row" fxLayoutAlign="space-evenly center">
+            <!-- <mat-slider (input)="onSliderChange($event)" step="0.005" min="0" max="100" class="slider">
+                <input matSliderThumb [(ngModel)]="sliderValue"/>
+            </mat-slider> -->
+            <button mat-raised-button (click)="openBottomSheet()">open</button>
+        </div>
     </div>
 </div>
